add chisel scripts
.m is just for testing equations
.py does the actual work of computing l based on t, h, r, a.

This model is useful for grinder stone sharpening:
- roughing gouge (45 deg)
- parting tool (40-55 deg)

The chisel is supposed to go into a jig that keeps the chisel
base at constant distance l from the axis of the grinding stone.
This way, one can sharpen a parting tool by moving ithe tip left and
right or a roughing gouge by rotating it left and right in the jig.